#622 - fixed icon handling for new items

+++ src/main/java/de/deadlocker8/budgetmaster/icon/Iconizable.java
@@ -16,11 +16,14 @@ public interface Iconizable extends ProvidesID
 		// the reference to an existing icon may already be null in the item that should be updated
 		// (e.g. user deletes the icon for the item and hit the save button)
 		// therefore the existing item must be retrieved from database by id
-		final Optional<T> existingItemOptional = itemService.findById(this.getID());
-		if(existingItemOptional.isPresent())
+		if(this.getID() != null)
 		{
-			final Iconizable existingItem = existingItemOptional.get();
-			iconService.deleteIcon(existingItem.getIconReference());
+			final Optional<T> existingItemOptional = itemService.findById(this.getID());
+			if(existingItemOptional.isPresent())
+			{
+				final Iconizable existingItem = existingItemOptional.get();
+				iconService.deleteIcon(existingItem.getIconReference());
+			}
 		}
 
 		final Optional<Icon> iconOptional = iconService.createIconReference(iconImageID, builtinIconIdentifier);
